Fix disemvowel
Return a String and not a reference to a str, as well as treating
disemvowel_ vec as a string, not a vector (but we forgot to rename
it). And instead of reffering to a chara reference to push it in
to the disemvowel vector, we can just return the String as it
was already pushed into in the for loop.

Basically, we didn't need to borrow in most places.

Thanks for tips from Karsh.